+++ Assets/200_Scripts/Characters/PlayerCharacter.cs
@@ -4,13 +4,13 @@ using UnityEngine;
 
 public class PlayerCharacter : Character
 {
-    public override void TakeDamage(float damage)
+    public override void TakeDamage(int damage)
     {
         base.TakeDamage(damage); // Appel à la méthode de la classe de base pour gérer les points de vie
         // Autres actions spécifiques au joueur
     }
 
-    public override void Heal(float amount)
+    public override void Heal(int amount)
     {
         // Logique spécifique pour le personnage du joueur lorsqu'il est soigné
         base.Heal(amount); // Appel à la méthode de la classe de base pour gérer les points de vie
